+++ DawnLib.Dusk/src/API/Definitions/Dungeon/DungeonConfig.cs
@@ -1,6 +1,9 @@
+using BepInEx.Configuration;
+
 namespace Dusk;
 
 public class DungeonConfig
 {
-
+    public ConfigEntry<string>? MoonSpawnWeights;
+    public ConfigEntry<string>? WeatherSpawnWeights;
 }